added dropdown component to favourites page
