+++ src/App.jsx
@@ -7,7 +7,6 @@ import Enter from "./pages/Enter/Enter";
 import SignUp from "./pages/SignUp/SignUp";
 import Contact from "./pages/Contact/Contact";
 import SignIn from "./pages/SignIn/SignIn";
-import DropDown from "./components/DropDown/DropDown";
 
 const App = () => {
   return (
@@ -19,7 +18,6 @@ const App = () => {
       <Route path="/contact" element={<Contact />} />
       <Route path="/favourites" element={<Favourites />} />
       <Route path="/about" element={<About />} />
-      <Route path="/DropDown" element={<DropDown />} />
     </Routes>
   );
 };
