+++ src/components/DropDown/DropDown.jsx
@@ -1,40 +1,25 @@
 import "./DropDown.scss";
 import triangle from "../../assets/icons/triangleUpsideDown.svg";
-
 import { useState } from "react";
 
 const DropDown = ({ options = [], onChange }) => {
-  const optionsArr = [
-    "All",
-    "Birmingham",
-    "London",
-    "Chester",
-    "Oxford",
-    "Newcastle",
-    "Liverpool",
-  ];
-
-  // you should call onChange with the new value whenever the value changes
-
-  // state to change whether drop down is displayed or not - changes class of image to rotate the arrow and class of primary div to change borders
-  const [dropdownActive, setDropdownActive] = useState(false);
-  // state to track the active location selected - This defaults to all so that without any option being clicked all is displayed as the text
+  const [dropDownActive, setDropDownActive] = useState(false);
   const [activeLocation, setActiveLocation] = useState("All");
 
   const handleDropDownActive = () => {
-    console.log("hi");
-    setDropdownActive(!dropdownActive);
+    setDropDownActive(!dropDownActive);
   };
+
   const handleActiveLocation = (location) => {
-    console.log(location);
+    onChange(location);
     setActiveLocation(`${location}`);
   };
 
-  const applyActiveClassDiv = dropdownActive
+  const applyActiveClassDiv = dropDownActive
     ? "dropdown__item dropdown__item--primary dropdown__item--primary--active"
     : "dropdown__item dropdown__item--primary";
 
-  const applyActiveClassArrow = dropdownActive
+  const applyActiveClassArrow = dropDownActive
     ? "dropdown__arrow arrow-active"
     : "dropdown__arrow";
 
@@ -50,11 +35,15 @@ const DropDown = ({ options = [], onChange }) => {
           tabIndex={0}
           onClick={handleDropDownActive}
         >
-          <img className={applyActiveClassArrow} src={triangle} alt="" />
+          <img
+            className={applyActiveClassArrow}
+            src={triangle}
+            alt="triangle"
+          />
         </div>
       </div>
-      {dropdownActive &&
-        optionsArr.map((location) => (
+      {dropDownActive &&
+        options.map((location) => (
           <div
             key={location}
             role="option"
@@ -70,14 +59,3 @@ const DropDown = ({ options = [], onChange }) => {
 };
 
 export default DropDown;
-
-//steps to do next
-
-// Filter locaiton results arry [1,2,3,4,45,5,67,]
-// user click on 3
-// 3 then is placed where all currently is, function the does resultsarray.push("all")
-// 4. if i then clicked 4- resultsarray.shift(3)
-// 4. if option clicked = all then remove all from results arry and put current value into array.
-
-// replace hardcoded shit with props.
-//  split the cuurrent component into container - default value shown ie all + an imported component with props -
