+++ src/components/DropDown/DropDown.scss
@@ -13,9 +13,12 @@
   flex-flow: column nowrap;
   justify-content: space-evenly;
   align-items: center;
-  border: 2px solid $black;
+  position: absolute;
+  padding: 1rem;
+
   .dropdown__item {
-    width: 100%;
+    background-color: $white;
+    min-width: 15rem;
     display: flex;
     flex-flow: row;
     justify-content: space-between;
